requirements.txt
# Django libararies
Django==4.1.4
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
rest-framework-generic-relations==2.1.0
psycopg2-binary==2.9.5
setuptools==65.6.0
gunicorn==20.1.0